NTC: use TS numbers in optimization problem (#449)
* NTC: use TS numbers

* Also use link TS numbers in adq-draft

* [sonar] Fix bug found by SonarCloud

* Set numSpace=0 in adq-draft and remove unused argument

* [sonar] Fix types to avoid loss of precision

+++ src/solver/simulation/adequacy-draft.cpp
@@ -81,9 +81,9 @@ bool AdequacyDraft::year(Progression::Task& progression,
 
     state.startANewYear();
 
-    for (unsigned int hour = pStartHour; hour != pFinalHour; ++hour)
+    for (uint hour = pStartHour; hour != pFinalHour; ++hour)
     {
-        state.hourInTheYear = (unsigned int)hour;
+        state.hourInTheYear = hour;
         state.eensSystemIS = 0.;
         state.eensSystemCN = 0.;
